optimize: unicode search

+++ source/include/sgl_core.h
@@ -280,10 +280,10 @@ typedef struct sgl_icon_pixmap {
 */
 typedef struct sgl_font_table {
     const uint32_t bitmap_index;
-    const int16_t box_h;
-    const int16_t box_w;
-    const int16_t ofs_x;
-    const int16_t ofs_y;
+    const int8_t box_h;
+    const int8_t box_w;
+    const int8_t ofs_x;
+    const int8_t ofs_y;
 
 }sgl_font_table_t;
